+++ src/Microsoft.Autofac/⬙/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ___
 
+## 2.1.0
+
+:calendar: _2023-03-25_
+
+### Added
+
+- A new overload of the `IRegistrationBuilder.WithLogger` extension method that allows to use a named logger that can be configured via a callback.
+___
+
 ## 2.0.0
 
 :calendar: _2022-12-30_